// utils/types.ts
export interface Gerations{
    id: number;
    name: string;
    url: string;
}